pyproject.toml
[tool.poetry]
name = "highlightreader"
version = "0.1.0"
description = ""
authors = ["Nickolai Yegorov <e-nicko@ya.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
requests = "^2.32.3"
pyjwt = "^2.9.0"
boto3 = "^1.35.29"
botocore = "^1.35.29"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
cryptography = "^43.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
